dataset/SyntheticDataGeneration/requirements.txt
transformers
datasets
sentence-transformers
faiss-cpu
snorkel
shap
uvicorn
fastapi
requests
pandas
pyarrow
boto3
